Conflicts:
	Cloudiversity.xcodeproj/project.pbxproj
	Cloudiversity/Main.storyboard
	Cloudiversity/Utils/CloudDateConverter.m
	Cloudiversity/Vendors/DSLCalendarView/DSLCalendarDayCalloutView.m
	Cloudiversity/Vendors/DSLCalendarView/DSLCalendarDayView.m
	Cloudiversity/Vendors/DSLCalendarView/DSLCalendarView.m
	Cloudiversity/ViewControllers/Abstracts/AbstractTableViewController.m
	Cloudiversity/ViewControllers/Abstracts/AbstractViewController.m
	Cloudiversity/ViewControllers/Agenda/AgendaFilterViewController.m
	Cloudiversity/ViewControllers/Agenda/AgendaStudentViewController.m
	Cloudiversity/ViewControllers/Agenda/AgendaTeacherClassViewController.m
	Cloudiversity/ViewControllers/Agenda/AgendaTeacherViewController.m
	Cloudiversity/ViewControllers/Navigation/NavigationViewController.m
	Cloudiversity/ViewControllers/Server/ServerViewController.m

+++ Cloudiversity/Utils/IOSRequest.h
@@ -36,17 +36,17 @@ typedef void (^HTTPFailureHandler)(AFHTTPRequestOperation *operation, NSError *e
 + (void)getAssignmentsForUserOnSuccess:(HTTPSuccessHandler)success
                              onFailure:(HTTPFailureHandler)failure;
 
-+ (void)getAssignmentsForClass:(int)classID
-                 andDiscipline:(int)disciplineID
++ (void)getAssignmentsForClass:(NSInteger)classID
+                 andDiscipline:(NSInteger)disciplineID
                      onSuccess:(HTTPSuccessHandler)success
                      onFailure:(HTTPFailureHandler)failure;
 
-+ (void)getAssignmentInformation:(int)assignmentId
++ (void)getAssignmentInformation:(NSInteger)assignmentId
                        onSuccess:(HTTPSuccessHandler)success
                        onFailure:(HTTPFailureHandler)failure;
 
-+ (void)updateAssignmentWithId:(int)assignmentId
-               withProgression:(int)progress
++ (void)updateAssignmentWithId:(NSInteger)assignmentId
+               withProgression:(NSInteger)progress
                      onSuccess:(HTTPSuccessHandler)success
                      onFailure:(HTTPFailureHandler)failure;
 
@@ -54,8 +54,8 @@ typedef void (^HTTPFailureHandler)(AFHTTPRequestOperation *operation, NSError *e
                     withDueDate:(NSString *)dueDate
                     withDueTime:(NSString *)dueTime
                 withDescription:(NSString *)description
-               withDisciplineID:(int)disciplineID
-                    withClassID:(int)classID
+               withDisciplineID:(NSInteger)disciplineID
+                    withClassID:(NSInteger)classID
                       onSuccess:(HTTPSuccessHandler)success
                       onFailure:(HTTPFailureHandler)failure;
 
@@ -63,9 +63,9 @@ typedef void (^HTTPFailureHandler)(AFHTTPRequestOperation *operation, NSError *e
                      withDueDate:(NSString *)dueDate
                      withDueTime:(NSString *)dueTime
                  withDescription:(NSString *)description
-                withDisciplineID:(int)disciplineID
-                     withClassID:(int)classID
-                 andAssignmentID:(int)assignmentID
+                withDisciplineID:(NSInteger)disciplineID
+                     withClassID:(NSInteger)classID
+                 andAssignmentID:(NSInteger)assignmentID
                        onSuccess:(HTTPSuccessHandler)success
                        onFailure:(HTTPFailureHandler)failure;
